# Fix Native SHA256 Performance Issue