# Fix Performance Issue by Eliminating AVX-to-SSE Transitions in Native SHA256 Implementation

## Description

This patch addresses a reported performance degradation caused by AVX-to-SSE transitions in the native SHA256 algorithm implementation. As seen in the following issue: https://github.com/FDio/vpp/issues/3579 The root cause was identified as the mixing of SHA instructions (which only have SSE variants) with AVX instructions such as `vpshufd` and `vpadd`, leading to costly transition penalties.

### Background

- SHA instructions do not have AVX variants and only operate using SSE.
- Combining these SSE SHA instructions with AVX instructions triggers AVX-to-SSE transitions.
- These transitions can be detected using Intel Performance Monitoring events (PMON) as described here: https://perfmon-events.intel.com/index.html?pltfrm=graniterapids_server.html&evnt=ASSISTS.SSE_AVX_MIX (Refresh page if it says 'Request Rejected')
- The recommended solution is to insert `vzeroupper` before the execution of the SSE SHA instructions to clear the upper portions of the YMM registers, preventing transition penalties.

### Changes made

- Inserted `vzeroupper` instructions appropriately before SSE SHA instruction blocks.
- Verified that this eliminates the costly transitions and improves performance.
- Confirmed no functional impact on the SHA256 implementation.

For detailed technical background, please refer to this document: https://www.intel.com/content/dam/develop/external/us/en/documents/11mc12-avoiding-2bavx-sse-2btransition-2bpenalties-2brh-2bfinal-809104.pdf

---

## Performance Results

| Metric                     | Before Patch | After Patch | 
|----------------------------|--------------|-------------|
| SHA256 Throughput (MB/s)   | 102         | 11,727  |     
| AVX-to-SSE Transitions     | 182,28,092         |  0   |

*Note:* Performance measured on INTEL(R) XEON(R) GOLD 6548N CPU 

---

Please review and let me know if further adjustments or benchmarks are needed.